Router authentication and characters

+++ src/app/components/games/game-detail/game-detail.component.ts
@@ -3,6 +3,7 @@ import { Component, OnInit} from '@angular/core';
 import { Game } from '../../../models/game.model';
 import { GameService } from '../../../services/game.service';
 import { ActivatedRoute,Router, Params } from '@angular/router';
+import { LoginService } from '../../../services/login.service'; 
 
 @Component({
   selector: 'app-game-detail',
@@ -14,7 +15,8 @@ export class GameDetailComponent implements OnInit {
 
   constructor(private gameService: GameService,
               private route: ActivatedRoute,
-              private router: Router
+              private router: Router,
+              private loginService: LoginService
   ) {}
 
   ngOnInit() {
@@ -29,14 +31,16 @@ export class GameDetailComponent implements OnInit {
         
       );
   }
-
-
   onEditGame(){
     this.router.navigate(['edit'],{relativeTo: this.route})
   }
 
   onDeleteGame(){
-    this.gameService.deleteGame(this.id);
-    this.router.navigate(['/games']);
+    if(this.loginService.currentUserValue){
+      this.gameService.deleteGame(this.id);
+      this.router.navigate(['/games']);
+    }else{
+      this.router.navigate(['/login']);
+    }
   }
 }
